pyproject.toml
# pyproject.toml - Python Package Configuration
# This file defines your project as a proper Python package

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "even_odd_league_player"
version = "0.1.0"
description = "AI Player Agent for Even/Odd League using MCP, FastAPI, and Agno+Gemini"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Lior Livyatan", email = "lior@example.com"}
]
keywords = ["llm", "agents", "ai", "mcp", "even-odd-league", "gemini", "agno"]

# Dependencies
dependencies = [
    # Data processing
    "numpy>=1.21.0",
    "pandas>=1.3.0",

    # HTTP Server & API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "httpx>=0.28.0",

    # AI & Agent Framework
    "agno>=0.59.0",
    "google-generativeai>=0.8.0",

    # Configuration & Utilities
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.2",
    "structlog>=24.4.0",
]

# Optional dependencies for development
[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",

    # Code Quality
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.7.0",

    # Analysis & Documentation
    "jupyter>=1.0.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.12.0",
    "ipython>=8.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/my_project"  # TODO: Update URL
Documentation = "https://github.com/yourusername/my_project/docs"  # TODO: Update URL
Repository = "https://github.com/yourusername/my_project"  # TODO: Update URL

# Tool configurations
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src --cov-report=html --cov-report=term-missing"

[tool.black]
line-length = 100
target-version = ['py38']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
